Add support for configuring hub settings through the client

+++ Client/ClientSettings.cs
@@ -36,7 +36,7 @@
             }
             set
             {
-                SettingsHelper.SetProperty(value);
+                SettingsHelper.SetProperty((int)value);
             }
         }
